+++ src/shared/ui/sections/contact/Information.tsx
@@ -1,0 +1,15 @@
+const Information = () => {
+    return (
+        <div className="rounded-lg bg-brand-primary px-52 py-16">
+            <div className="max-w-3xl mx-auto text-center">
+                <h1 className="text-5xl font-bold text-black mb-6">Контакты</h1>
+                <p className="text-xl text-black">
+                    Мы всегда рады помочь вам с выбором продукта или ответить на любые вопросы. Свяжитесь с нами удобным для вас
+                    способом.
+                </p>
+            </div>
+        </div>
+    )
+}
+
+export default Information;
